Rename grid_container file
