+++ tests/conftest.py
@@ -5,7 +5,7 @@ import pytest
 
 from definitions import ROOT_DIR
 from psdm_analysis.models.grid_with_results import GridWithResults
-from psdm_analysis.models.input.container.grid_container import GridContainer
+from psdm_analysis.models.input.container.grid import GridContainer
 
 # Note: the fixture scope defines how long the fixture is valid and prevents reevaluation
